Pull out Abi class for ABI encoding/decoding
