+++ packages/iov-ethereum/types/encoding.d.ts
@@ -20,15 +20,3 @@ export declare type Eip155ChainId = {
 };
 export declare function eip155V(chain: Eip155ChainId, recoveryParam: number): number;
 export declare function getRecoveryParam(chain: Eip155ChainId, v: number): number;
-export interface HeadTail {
-    /** An array of start positions within the original data */
-    readonly head: ReadonlyArray<number>;
-    /** Arguments split by positions as defined by head */
-    readonly tail: ReadonlyArray<Uint8Array>;
-}
-/**
- * Decode head-tail encoded data as described in
- * https://medium.com/@hayeah/how-to-decipher-a-smart-contract-method-call-8ee980311603
- */
-export declare function decodeHeadTail(data: Uint8Array): HeadTail;
-export declare function decodeVariableLength(data: Uint8Array): Uint8Array;
